requirements.txt
aiofiles==23.2.1  
annotated-types==0.7.0  
anyio==4.8.0  
certifi==2024.12.14  
charset-normalizer==3.4.1  
click==8.1.8  
distro==1.9.0  
elevenlabs==1.50.3  
fastapi==0.115.6  
ffmpy==0.5.0  
filelock==3.16.1  
fsspec==2024.12.0  
gradio==5.12.0  
gradio-client==1.5.4  
groq==0.15.0  
gtts==2.5.4  
h11==0.14.0  
httpcore==1.0.7  
httpx==0.28.1  
huggingface-hub==0.27.1  
idna==3.10  
jinja2==3.1.5  
markdown-it-py==3.0.0  
markupsafe==2.1.5  
mdurl==0.1.2  
numpy==2.2.1  
orjson==3.10.14  
packaging==24.2  
pandas==2.2.3  
pillow==11.1.0  
pyaudio==0.2.14  
pydantic==2.10.5  
pydantic-core==2.27.2  
pydub==0.25.1  
pygments==2.19.1  
python-dateutil==2.9.0.post0  
python-multipart==0.0.20  
pytz==2024.2  
pyyaml==6.0.2  
requests==2.32.3  
rich==13.9.4  
ruff==0.9.1  
safehttpx==0.1.6  
semantic-version==2.10.0  
shellingham==1.5.4  
six==1.17.0  
sniffio==1.3.1  
speechrecognition==3.13.0  
starlette==0.41.3  
tomlkit==0.13.2  
tqdm==4.67.1  
typer==0.15.1  
typing-extensions==4.12.2  
tzdata==2024.2  
urllib3==2.3.0  
uvicorn==0.34.0  
websockets==14.1  
python-dotenv  
pydub